tests: add --keep-build-dir option to prevent build directory from being removed
Useful when trying to debug what went wrong.

Changelog: title

(cherry picked from commit 414521b51878bcb598749f2a71e2653d3f743c51)

+++ tests/acceptance/common.py
@@ -491,7 +491,8 @@ def prepared_test_build_base(request, bitbake_variables):
     build_dir = tempfile.mkdtemp(prefix="test-build-", dir=os.environ['BUILDDIR'])
 
     def cleanup_test_build():
-        run_verbose("rm -rf %s" % build_dir)
+        if not pytest.config.getoption('--keep-build-dir'):
+            run_verbose("rm -rf %s" % build_dir)
 
     cleanup_test_build()
     request.addfinalizer(cleanup_test_build)
